+++ app/main.py
@@ -1,10 +1,15 @@
 from src.logger.logging import setup_logging
 from src.tools.utils import parse_arguments
 
+from app.src.content.site_content import SiteContent
+
 
 def main():
     args = parse_arguments()
     setup_logging(log_level=args.log_level, log_level_others=args.log_level_other)
+
+    site_content = SiteContent()
+    site_content.get_content()
 
 
 if __name__ == "__main__":