+++ CHANGELOG.md
@@ -17,3 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Argument Parser
 - Isort and Black configuration
 - Loading environment values
+
+## 2026-02-19
+### Added
+- Fetching HTML content from URL to file